requirements.txt
# Core Dependencies
psycopg2-binary==2.9.9
groq==1.2.0
python-dotenv==1.0.0
PyYAML==6.0.1
pydantic==2.12.5
pydantic-settings==2.3.4
redis==7.1.0
httpx==0.28.1

# Web Dashboard (Flask — kept in parallel during FastAPI migration)
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0

# Async API (FastAPI)
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# Background Tasks
celery[redis]==5.4.0

# GitHub/GitLab Integration
PyGithub==2.1.1
python-gitlab==4.4.0

# CLI & UI
rich==13.7.0

# Monitoring
prometheus-client==0.21.1

# Static Analysis Tools (pinned to prevent rule ID drift)
ruff==0.6.0
semgrep==1.45.0
vulture==2.10
radon==6.0.1
bandit==1.7.5

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Database Migrations
alembic==1.13.1
sqlalchemy==2.0.23

# Utilities
tenacity==8.2.3
requests==2.31.0
